pyproject.toml
[project]
name = "agent-a-to-z"
version = "0.1.0"
description = "AI Agent with crewai"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai[tools]>=0.22.0",
    "dotenv>=0.9.9",
    "python-telegram-bot[job-queue]>=22.3",
    "firecrawl-py>=3.4.0",
    "feedparser>=6.0.11",
    "langchain-openai>=0.3.35",
    "langchain>=0.3.27",
    "pytz>=2025.2",
]
